requirements.txt
# Data Processing
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2

# Machine Learning
xgboost==2.0.2
lightgbm==4.1.0
catboost==1.2.2

# MLOps Tools
mlflow==2.9.1
dvc==3.36.1
great-expectations==0.18.5

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Code Quality
black==23.12.0
flake8==6.1.0
mypy==1.7.1
isort==5.13.2

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
click==8.1.7
tqdm==4.66.1
joblib==1.3.2

# API & Deployment
fastapi==0.108.0
uvicorn==0.25.0
pydantic==2.5.2

# Jupyter
jupyter==1.0.0
ipykernel==6.27.1
